focus on evalution
-----------todo-----------
1. proof
2. average score
3. integrate limt role

+++ resources/views/accompaniment/index.blade.php
@@ -7,37 +7,36 @@
 </style>
 <link href="{{ asset('css/fab.css') }}" rel="stylesheet">
 @foreach($data as $key=>$value)
-<div class="container-fluid flex-wrap">
+@foreach($isMember as $iM)
+@if($iM == Auth::id())
+<div class="container-fluid flex-wrap" style="padding: 2%;">
     @foreach ($value as $val)
     <div id="s{{ $val->classroom }}">
-    @endforeach
+        @endforeach
         <div class="accordion-item">
             <h5 id="headingOne" class="accordion-header">
-                <button class="accordion-button collapsed" data-bs-toggle="collapse"
-                    data-bs-target="#s{{ $val->classroom }}One">
+                <button class="accordion-button collapsed" data-bs-toggle="collapse" data-bs-target="#s{{ $val->classroom }}One">
                     {{ $key }}
                 </button>
             </h5>
         </div>
         <div class="card">
             @foreach ($value as $val)
-            <div id="s{{ $val->classroom }}One" class="accordion-collapse collapse" aria-labelledby="headingOne"
-                data-bs-parent="#s{{ $val->classroom }}">
+            <div id="s{{ $val->classroom }}One" class="accordion-collapse collapse" aria-labelledby="headingOne" data-bs-parent="#s{{ $val->classroom }}">
                 <div class=" d-flex bd-highlight">
-                    <button class="accordion-button p-2 flex-grow-1 bd-highlight" data-bs-toggle="collapse"
-                        data-bs-target="#s{{ $val->classroom }}OneA">{{ $val->title }}
+                    <button class="accordion-button p-2 flex-grow-1 bd-highlight" data-bs-toggle="collapse" data-bs-target="#s{{ $val->classroom }}OneA">{{ $val->title }}
                     </button>
                     <a href="../accompaniment/create/{{ $val->id }}">
                         <i class="fa fa-edit fa-3x imgBtn" id="edit_note" style="padding: 5px;"></i>
                     </a>
-                    <a href="../accompaniment/delete/{{ $val->id }}">
+                    <a href="../accompaniment/delete/{{ $val->id }}" onclick="return confirm('Are you sure?');">
                         <i class="fa fa-trash fa-3x imgBtn" id="delete_note" style="padding: 5px;"></i>
                     </a>
                 </div>
-                <div class="accordion-collapse collapse" id="s{{ $val->classroom }}OneA">
-                    <div class="accordion-body" style="text-overflow: ellipsis; width: 50ch; overflow: hidden;
+                <div class="bg-info accordion-collapse collapse" id="s{{ $val->classroom }}OneA">
+                    <div class="accordion-body" style="text-overflow: ellipsis; width: 100%; overflow: hidden;
                             white-space: nowrap;">
-                        {{ $val->description }}
+                        {{ strip_tags( $val->description ) }}
                     </div>
                 </div>
             </div>
@@ -45,8 +44,12 @@
         </div>
     </div>
 </div>
+@endif
 @endforeach
+@endforeach
+@if($authRole->roleName!='mahasiswa')
 <a href="../accompaniment/create/0" class="float">
     <i class="fa fa-plus my-float"></i>
 </a>
+@endif
 @endsection
